Move TestDestroyCache to main DataStreamerTest
