+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Dataload/DataStreamerTest.cs
@@ -588,6 +588,64 @@ namespace Apache.Ignite.Core.Tests.Dataload
         }
 
         /// <summary>
+        /// Streamer test with destroyed cache.
+        /// </summary>
+        [Test]
+        public void TestDestroyCache()
+        {
+            var cache = _grid.CreateCache<int, int>(TestUtils.TestName);
+
+            var streamer = _grid.GetDataStreamer<int, int>(cache.Name);
+
+            streamer.Add(1, 2);
+            var task = streamer.BatchTask;
+            streamer.Flush();
+            task.Wait();
+
+            _grid.DestroyCache(cache.Name);
+
+            streamer.Add(2, 3);
+            task = streamer.BatchTask;
+            streamer.Flush();
+
+            var ex = Assert.Throws<AggregateException>(task.Wait).InnerException;
+
+            Assert.IsNotNull(ex);
+
+            Assert.AreEqual("class org.apache.ignite.IgniteCheckedException: DataStreamer data loading failed.",
+                ex.Message);
+        }
+
+        /// <summary>
+        /// Streamer test with destroyed cache.
+        /// </summary>
+        [Test]
+        public void TestDestroyCacheObsolete()
+        {
+#pragma warning disable 618
+            var cache = _grid.CreateCache<int, int>(TestUtils.TestName);
+
+            var streamer = _grid.GetDataStreamer<int, int>(cache.Name);
+
+            var task = streamer.AddData(1, 2);
+            streamer.Flush();
+            task.Wait();
+
+            _grid.DestroyCache(cache.Name);
+
+            task = streamer.AddData(2, 3);
+            streamer.Flush();
+
+            var ex = Assert.Throws<AggregateException>(task.Wait).InnerException;
+
+            Assert.IsNotNull(ex);
+
+            Assert.AreEqual("class org.apache.ignite.IgniteCheckedException: DataStreamer data loading failed.",
+                ex.Message);
+#pragma warning restore 618
+        }
+
+        /// <summary>
         /// Test binarizable receiver.
         /// </summary>
         private class StreamReceiverBinarizable : IStreamReceiver<int, int>
