+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Dataload/DataStreamerTestTopologyChange.cs
@@ -68,37 +68,5 @@ namespace Apache.Ignite.Core.Tests.Dataload
                                 "(all affinity nodes have left the grid or cache was stopped): cache]", ex.Message);
             }
         }
-
-        /// <summary>
-        /// Streamer test with destroyed cache.
-        /// </summary>
-        [Test]
-        public void TestDestroyCache()
-        {
-            const string cacheName = "cache";
-
-            using (var grid = Ignition.Start(TestUtils.GetTestConfiguration()))
-            {
-                grid.CreateCache<int, int>(cacheName);
-
-                var streamer = grid.GetDataStreamer<int, int>(cacheName);
-
-                var task = streamer.AddData(1, 2);
-                streamer.Flush();
-                task.Wait();
-
-                grid.DestroyCache(cacheName);
-
-                task = streamer.AddData(2, 3);
-                streamer.Flush();
-
-                var ex = Assert.Throws<AggregateException>(task.Wait).InnerException;
-
-                Assert.IsNotNull(ex);
-
-                Assert.AreEqual("class org.apache.ignite.IgniteCheckedException: DataStreamer data loading failed.", 
-                    ex.Message);
-            }
-        }
     }
 }
